requirements.txt
selene==2.0.0a40
pytest==6.2.4
uuid==1.30
pytest-testdox==2.0.1
tenacity==7.0.0
pytest-env==0.6.2
pytest-rerunfailures==9.1.1
faker==8.1.0
allure-pytest==2.8.40
pytest-xdist==2.2.1
selenium-wire==4.6.4
pre-commit==2.19.0


selenium~=3.141.0